[IR] Moved `compiler.lang.backend.Caching` to `compiler.opt.Caching`.
